Check if the files exist in the Pipeline
This commit adds file checks to the Pipeline class. This means now
Pipeline will throw a FileNotFoundException instance to indicate that
any file, specified with the !files keyword, is missing.
