+++ tests/CMakeLists.txt
@@ -44,6 +44,7 @@ link_libraries(
 )
 
 add_subdirectory(application)
+add_subdirectory(exceptions)
 add_subdirectory(lexer)
 add_subdirectory(parser)
 add_subdirectory(scheduler)
