+++ tests/scheduler/pipeline/pipeline/CMakeLists.txt
@@ -24,6 +24,7 @@ set(SOURCES
 )
 
 set(STUBS
+    ${STUBS_FOLDER}/exceptions/filenotfoundexception.cpp
     ${STUBS_FOLDER}/scheduler/exceptions/compilationerrorexception.cpp
     ${STUBS_FOLDER}/scheduler/exceptions/linkerrorexception.cpp
     ${STUBS_FOLDER}/scheduler/exceptions/nofilesspecifiedexception.cpp
